move palette switch / shortcut setup from core to blend menu
