+++ src/components/core/mineral-controls.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, ReactElement } from 'react'
+import { useState, ReactElement } from 'react'
 import { MdColorLens } from 'react-icons/md'
 import { useBlendState, useBlending } from '../../hooks/blend-context'
 import { GenericPalette } from '../../lib/palettes'
@@ -17,43 +17,11 @@ function CoreMineralControls (
 ): ReactElement {
     const [menuOpen, setMenuOpen] = useState<boolean>(false)
     const {
-        palette,
         visibilities,
         setVisibilities,
-        monochrome,
         setMonochrome
     } = useBlendState()
     useBlending(vis)
-
-    // setup keyboard shortcuts
-    useEffect(() => {
-        const keydown = (e: KeyboardEvent): void => {
-            if (e.key === 'b') {
-                setMonochrome(!monochrome)
-                return
-            }
-            const numKey = parseInt(e.key)
-            if (numKey > 0 && numKey <= minerals.length) {
-                visibilities[numKey - 1] = !visibilities[numKey - 1]
-                setVisibilities([...visibilities])
-            }
-        }
-        window.addEventListener('keydown', keydown)
-        return () => {
-            window.removeEventListener('keydown', keydown)
-        }
-    }, [monochrome, visibilities, minerals, setMonochrome, setVisibilities])
-
-    // update visibilities to match newly selected palette on change
-    useEffect(() => {
-        if (palette.type === 'labelled') {
-            const visibleMinerals = Object.keys(palette.colors)
-            setVisibilities(minerals.map(mineral => visibleMinerals.includes(mineral)))
-        } else {
-            const numVisible = palette.colors.length
-            setVisibilities(minerals.map((_, i) => i < numVisible))
-        }
-    }, [palette, minerals, setVisibilities])
 
     // sets parameters to show one channel in monochrome
     const getMineralSetter = (i: number): (() => void) => {
